+++ src/steps/constants.ts
@@ -10,13 +10,24 @@ export const Steps = {
   USERS: 'fetch-users',
   ACCOUNT: 'fetch-account',
   ROLES: 'fetch-roles',
+  HOSTS: 'fetch-hosts',
   BUILD_ACCOUNT_USER_RELATIONSHIPS: 'build-account-user-relationships',
   BUILD_ACCOUNT_ROLE_RELATIONSHIPS: 'build-account-role-relationships',
   BUILD_USER_ROLE_RELATIONSHIPS: 'build-user-role-relationships',
 };
 
-export const Entities: Record<'USER' | 'ACCOUNT' | 'ROLE', StepEntityMetadata> =
+export const Entities: Record<
+  | 'HOST'
+  | 'USER'
+  | 'ACCOUNT'
+  | 'ROLE',
+  StepEntityMetadata> =
   {
+    HOST: {
+      resourceName: 'Host',
+      _type: 'datadog_host',
+      _class: ['Host'],
+    },
     USER: {
       resourceName: 'User',
       _type: 'datadog_user',
@@ -35,9 +46,18 @@ export const Entities: Record<'USER' | 'ACCOUNT' | 'ROLE', StepEntityMetadata> =
   };
 
 export const Relationships: Record<
-  'ACCOUNT_HAS_USER' | 'ACCOUNT_HAS_ROLE' | 'USER_ASSIGNED_ROLE',
+  | 'ACCOUNT_HAS_HOST'
+  | 'ACCOUNT_HAS_USER'
+  | 'ACCOUNT_HAS_ROLE'
+  | 'USER_ASSIGNED_ROLE',
   StepRelationshipMetadata
 > = {
+  ACCOUNT_HAS_HOST: {
+    _type: 'datadog_account_has_host',
+    _class: RelationshipClass.HAS,
+    sourceType: Entities.ACCOUNT._type,
+    targetType: Entities.HOST._type,
+  },
   ACCOUNT_HAS_USER: {
     _type: 'datadog_account_has_user',
     _class: RelationshipClass.HAS,